feat(hoist-field): refactor to enable bare implementation
