+++ packages/transforms/hoist-field/src/bareHoistField.ts
@@ -1,0 +1,9 @@
+import { MeshTransform, MeshTransformOptions, YamlConfig } from '@graphql-mesh/types';
+
+export default class BareHoistField implements MeshTransform {
+  noWrap = true;
+
+  constructor(options: MeshTransformOptions<YamlConfig.HoistFieldTransformConfig>) {
+    throw new Error('Not implemented');
+  }
+}
